+++ protocols/substreams/ethereum-ekubo-v3/src/abi/mod.rs
@@ -2,3 +2,4 @@
 pub mod boosted_fees;
 pub mod core;
 pub mod twamm;
+pub mod ve33;